workshop-201/teachers_assistant/pyproject.toml
[project]
name = "teachers-assistant"
version = "0.1.0"
description = "Teaching Assistant Workshop"
requires-python = ">=3.13"
dependencies = [
    "langchain==0.3.27",
    "langchain-community==0.3.30",
    "langchain-google-vertexai==2.1.2",
    "matplotlib==3.10.6",
    "ollama==0.6.0",
    "ragas==0.3.5",
    "requests==2.32.5",
    "seaborn==0.13.2",
    "strands-agents==1.10.0",
    "strands-agents-tools==0.2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"